+++ src/settings/StaffSlips/tokens.test.js
@@ -19,6 +19,11 @@ describe('getTokens', () => {
   const fieldsForCheck = [
     {
       category: 'item',
+      name: 'item.displaySummary',
+      expectedResult: 'no.1-2v. 121948-1962 (Board)',
+    },
+    {
+      category: 'item',
       name: 'item.lastCheckedInDateTime',
       expectedResult: expectedDateResult,
     },