+++ src/settings/LoanPolicy/LoanPolicyForm.test.js
@@ -193,8 +193,7 @@ describe('LoanPolicyForm', () => {
 
   it('should execute "FooterPane" with passed props', () => {
     expect(FooterPane).toHaveBeenCalledWith({
-      pristine: true,
-      submitting: false,
+      isSaveButtonDisabled: true,
       onCancel: mockedOnCancel,
     }, {});
   });