+++ src/settings/LostItemFeePolicy/LostItemFeePolicyForm.test.js
@@ -110,8 +110,7 @@ describe('LostItemFeePolicyForm', () => {
 
   it('should execute "FooterPane" with passed props', () => {
     expect(FooterPane).toHaveBeenCalledWith({
-      pristine: true,
-      submitting: false,
+      isSaveButtonDisabled: true,
       onCancel: mockedOnCancel,
     }, {});
   });