+++ src/settings/NoticePolicy/NoticePolicyForm.test.js
@@ -146,8 +146,7 @@ describe('NoticePolicyForm', () => {
 
   it('should execute "FooterPane" with passed props', () => {
     expect(FooterPane).toHaveBeenCalledWith({
-      pristine: true,
-      submitting: false,
+      isSaveButtonDisabled: true,
       onCancel: mockedOnCancel,
     }, {});
   });