+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change history for ui-circulation
 
+## 9.0.5 IN PROGRESS
+
+* Only certain HTML tags should be rendered when displaying staff slips. Refs UICIRC-1070.
+
 ## [9.0.4](https://github.com/folio-org/ui-circulation/tree/v9.0.4) (2024-02-22)
 [Full Changelog](https://github.com/folio-org/ui-circulation/compare/v9.0.3...v9.0.4)